Fix download overwriting and improve messages
