--- data/train.txt
@@ -1,3 +1,0 @@
-https://www.youtube.com/watch?v=l1hMkDTg2lg
-https://www.youtube.com/playlist?list=PLC1qU-LWwrF64f4QKQT-Vg5Wr4qEE1Zxk
-https://www.youtube.com/playlist?list=PLUl4u3cNGP63WbdFxL8giv4yhgdMGaZNA
